Split modules for api and parsing templates
